use concerns to DRY ProtocolsController
