+++ app/assets/stylesheets/custom/bootstrap-table-custom.scss
@@ -43,7 +43,11 @@
       padding: 0 1rem;
 
       &> * {
-        margin: 1rem 0 1rem 5px;
+        margin: 1rem 0 1rem 0px;
+      }
+
+      .columns {
+        margin-left: 5px;
       }
     }
